refactor: links

+++ src/main/java/com/backend/quack/service/LinkService.java
@@ -1,6 +1,7 @@
 package com.backend.quack.service;
 
 import com.backend.quack.domain.entity.Collection;
+import com.backend.quack.domain.entity.CollectionVisibility;
 import com.backend.quack.domain.entity.Link;
 import com.backend.quack.domain.repository.CollectionRepository;
 import com.backend.quack.domain.repository.LinkRepository;
@@ -10,6 +11,8 @@ import com.backend.quack.dto.link.LinkResponseDTO;
 import com.backend.quack.exception.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,23 +24,55 @@ public class LinkService {
     private final LinkRepository linkRepository;
     private final CollectionRepository collectionRepository;
 
-    public List<LinkResponseDTO> listAllLinksInCollection(long collectionId) {
-        List<Link> links = this.linkRepository.findAllByCollectionIdAndDeletedIsFalse(collectionId);
-        return links
-                .stream()
-                .map(LinkResponseDTO::fromEntity)
-                .toList();
+    public Page<LinkResponseDTO> listAllLinksInCollection(String username, String slug, Pageable pageable) {
+        Collection collection = collectionRepository.findBySlug(slug);
+
+        if (collection == null) {
+            throw new EntityNotFoundException("Collection not found");
+        }
+
+        if (collection.getVisibility().equals(CollectionVisibility.PRIVATE)) {
+            if (username == null) throw new EntityNotFoundException("Collection not found");
+
+            if (!collection.getUser().getUsername().equals(username)) {
+                throw new EntityNotFoundException("Collection not found");
+            }
+        }
+
+        Page<Link> links = this.linkRepository.findAllByCollectionSlugAndDeletedIsFalse(slug, pageable);
+        return links.map(LinkResponseDTO::fromEntity);
     }
 
-    public LinkResponseDTO findLinkById(long id) {
-        Link link = this.linkRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Link not found"));
+    public LinkResponseDTO findLinkById(String username, String slug, long id) {
+        Collection collection = collectionRepository.findBySlug(slug);
+
+        if (collection == null) {
+            throw new EntityNotFoundException("Collection not found");
+        }
+
+        if (collection.getVisibility().equals(CollectionVisibility.PRIVATE)) {
+            if (username == null) throw new EntityNotFoundException("Collection not found");
+
+            if (!collection.getUser().getUsername().equals(username)) {
+                throw new EntityNotFoundException("Collection not found");
+            }
+        }
+
+        Link link = this.linkRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Link not found"));
         return LinkResponseDTO.fromEntity(link);
     }
 
-    public LinkResponseDTO saveLink(LinkPostDTO linkPostDTO) {
-        Collection collection = collectionRepository
-                .findByIdAndDeletedIsFalse(linkPostDTO.collection_id())
-                .orElseThrow(() -> new EntityNotFoundException("Collection not Found"));
+    public LinkResponseDTO saveLink(String username, String slug, LinkPostDTO linkPostDTO) {
+        Collection collection = collectionRepository.findBySlug(slug);
+
+        if (collection == null) {
+            throw new EntityNotFoundException("Collection not found");
+        }
+
+        if (!collection.getUser().getUsername().equals(username)) {
+            throw new EntityNotFoundException("Collection not found");
+        }
 
         Link link = linkPostDTO.toEntity();
         link.setCollection(collection);
@@ -45,14 +80,24 @@ public class LinkService {
         return LinkResponseDTO.fromEntity(this.linkRepository.save(link));
     }
 
-    public LinkResponseDTO updateLinkById(long id, LinkPutDTO linkPutDTO) {
+    public LinkResponseDTO updateLinkById(String username, long id, LinkPutDTO linkPutDTO) {
         Link link = this.linkRepository.getReferenceById(id);
+
+        if (!link.getCollection().getUser().getUsername().equals(username)) {
+            throw new EntityNotFoundException("Link not found");
+        }
+
         link.update(linkPutDTO.toEntity());
         return LinkResponseDTO.fromEntity(link);
     }
 
-    public void deleteLinkById(long id) {
+    public void deleteLinkById(long id, String username) {
         Link link = this.linkRepository.getReferenceById(id);
+
+        if (!link.getCollection().getUser().getUsername().equals(username)) {
+            throw new EntityNotFoundException("Link not found");
+        }
+
         link.delete();
     }
 }
